Conflicts:
	java/com.sap.sailing.gwt.ui/META-INF/MANIFEST.MF
	java/com.sap.sailing.server/SailingServer (No Proxy).launch
	java/com.sap.sailing.server/SailingServer (Proxy).launch
	java/com.sap.sailing.targetplatform/definitions/race-analysis-p2-ide-local.target

+++ configuration/buildAndUpdateProduct.sh
@@ -61,16 +61,19 @@ HAS_OVERWRITTEN_TARGET=0
 TARGET_SERVER_NAME=$active_branch
 
 gwtcompile=1
+onegwtpermutationonly=0
 testing=1
 clean="clean"
 offline=0
 proxy=0
+no_confirmation=1
 extra=''
 
 if [ $# -eq 0 ]; then
-    echo "buildAndUpdateProduct [-g -t -o -c -m <config> -n <package> -l <port>] [build|install|all|hot-deploy|remote-deploy]"
+    echo "buildAndUpdateProduct [-b -u -g -t -o -c -m <config> -n <package> -l <port>] [build|install|all|hot-deploy|remote-deploy]"
     echo ""
     echo "-g Disable GWT compile, no gwt files will be generated, old ones will be preserved."
+    echo "-b Build GWT permutation only for one browser and English language."
     echo "-t Disable tests"
     echo "-o Enable offline mode (does not work for tycho surefire plugin)"
     echo "-c Disable cleaning (use only if you are sure that no java file has changed)"
@@ -81,10 +84,11 @@ if [ $# -eq 0 ]; then
     echo "-l <telnet port>  Telnet port the OSGi server is running. Optional but enables fully automatic hot-deploy."
     echo "-s <target server> Name of server you want to use as target for install, hot-deploy or remote-reploy. This overrides default behaviour."
     echo "-w <ssh target> Target for remote-deploy. Must comply with the following format: user@server."
+    echo "-u Run without confirmation messages. Use with extreme care."
     echo ""
     echo "build: builds the server code using Maven to $PROJECT_HOME (log to $START_DIR/build.log)"
     echo "install: installs product and configuration to $SERVERS_HOME/$active_branch. Overwrites any configuration by using config from branch."
-    echo "all: calls build and then install"
+    echo "all: invokes build and then install"
     echo ""
     echo "hot-deploy: performs hot deployment of named bundle into OSGi server"
     echo "Example: $0 -n com.sap.sailing.www -l 14888 hot-deploy"
@@ -104,12 +108,13 @@ echo PROJECT_HOME is $PROJECT_HOME
 echo SERVERS_HOME is $SERVERS_HOME
 echo BRANCH is $active_branch
 
-options=':gtocpm:n:l:s:w:'
+options=':bgtocpm:n:l:s:w:u'
 while getopts $options option
 do
     case $option in
         g) gwtcompile=0;;
         t) testing=0;;
+	    b) onegwtpermutationonly=1;;
         o) offline=1;;
         c) clean="";;
         p) proxy=1;;
@@ -119,6 +124,7 @@ do
         s) TARGET_SERVER_NAME=$OPTARG
            HAS_OVERWRITTEN_TARGET=1;;
         w) REMOTE_SERVER_LOGIN=$OPTARG;;
+        u) no_confirmation=1;;
         \?) echo "Invalid option"
             exit 4;;
     esac
@@ -146,13 +152,13 @@ if [[ "$@" == "hot-deploy" ]]; then
         exit
     fi
 
+    if [[ $HAS_OVERWRITTEN_TARGET -eq 1 ]]; then
+        active_branch=$TARGET_SERVER_NAME
+    fi
+
     if [ ! -d $SERVERS_HOME/$active_branch/plugins ]; then
         echo "Could not find target directory $SERVERS_HOME/$active_branch/plugins!"
         exit
-    fi
-
-    if [[ $HAS_OVERWRITTEN_TARGET -eq 1 ]]; then
-        active_branch=$TARGET_SERVER_NAME
     fi
 
     # locate old bundle
@@ -179,12 +185,14 @@ if [[ "$@" == "hot-deploy" ]]; then
         echo "WARNING: Bundle versions do not differ. Update not needed."
     fi
 
-    read -s -n1 -p "Do you really want to hot-deploy bundle $OSGI_BUNDLE_NAME to $SERVERS_HOME/$active_branch? (y/N): " answer
-    case $answer in
-    "Y" | "y") echo "Continuing";;
-    *) echo "Aborting..."
-       exit;;
-    esac
+	if [ $no_confirmation -eq 0 ]; then
+        read -s -n1 -p "Do you really want to hot-deploy bundle $OSGI_BUNDLE_NAME to $SERVERS_HOME/$active_branch? (y/N): " answer
+        case $answer in
+        "Y" | "y") echo "Continuing";;
+        *) echo "Aborting..."
+           exit;;
+        esac
+    fi
 
     # deploy new bundle physically
     mkdir -p $SERVERS_HOME/$active_branch/plugins/deploy
@@ -223,12 +231,15 @@ if [[ "$@" == "hot-deploy" ]]; then
     BUNDLE_ID=`echo $OLD_BUNDLE_INFORMATION | cut -d " " -f 1`
     OLD_ACTIVATED_NAME=`echo $OLD_BUNDLE_INFORMATION | cut -d " " -f 3`
     echo "Could identify bundle-id $BUNDLE_ID for $OLD_ACTIVATED_NAME"
-    read -s -n1 -p "I will now stop and reinstall the bundle mentioned in the line above. Is this right? (y/N): " answer
-    case $answer in
-    "Y" | "y") echo "Continuing";;
-    *) echo "Aborting..."
-       exit;;
-    esac
+
+	if [ $no_confirmation -eq 0 ]; then
+        read -s -n1 -p "I will now stop and reinstall the bundle mentioned in the line above. Is this right? (y/N): " answer
+        case $answer in
+        "Y" | "y") echo "Continuing";;
+        *) echo "Aborting..."
+           exit;;
+        esac
+    fi
 
     # stop and uninstall
     echo -n stop $BUNDLE_ID | $NC_CMD > /dev/null
@@ -270,6 +281,24 @@ if [[ "$@" == "build" ]] || [[ "$@" == "all" ]]; then
 	if [ $gwtcompile -eq 1 ]; then
 	    echo "INFO: Compiling GWT (rm -rf com.sap.sailing.gwt.ui/com.sap.sailing.*)"
 	    rm -rf com.sap.sailing.gwt.ui/com.sap.sailing.*
+        if [ $onegwtpermutationonly -eq 1 ]; then
+            echo "INFO: Patching .gwt.xml files such that only one GWT permutation needs to be compiled"
+            for i in com.sap.sailing.gwt.ui/src/main/resources/com/sap/sailing/gwt/ui/*.gwt.xml; do
+                echo "INFO: Patching $i files such that only one GWT permutation needs to be compiled"
+                cp $i $i.bak
+                cat $i | sed -e 's/^[	 ]*<extend-property  *name="locale"  *values="de" *\/>/<!-- <extend-property name="locale" values="de"\/> --> <set-property name="user.agent" value="safari" \/>/' >$i.sed
+                mv $i.sed $i
+            done
+        else
+            echo "INFO: Patching .gwt.xml files such that all GWT permutations are compiled"
+            for i in com.sap.sailing.gwt.ui/src/main/resources/com/sap/sailing/gwt/ui/*.gwt.xml; do
+                echo "INFO: Patching $i files such that all GWT permutations are compiled"
+                cp $i $i.bak
+                cat $i | sed -e 's/<!-- <extend-property  *name="locale"  *values="de" *\/> --> <set-property name="user.agent" value="safari" \/>/<extend-property name="locale" values="de"\/>/' >$i.sed
+                mv $i.sed $i
+            done
+        fi
+
 	else
 	    echo "INFO: GWT Compilation disabled"
 	    extra="-Pdebug.no-gwt-compile"
@@ -298,17 +327,27 @@ if [[ "$@" == "build" ]] || [[ "$@" == "all" ]]; then
 	echo "Using following command: mvn $extra -fae -s $MAVEN_SETTINGS $clean install"
 	mvn $extra -fae -s $MAVEN_SETTINGS $clean install 2>&1 | tee $START_DIR/build.log
 
+	if [ $gwtcompile -eq 1 ]; then
+	    # Now move back the backup .gwt.xml files before they were (maybe) patched
+		echo "INFO: restoring backup copies of .gwt.xml files after they has been patched before"
+	    for i in com.sap.sailing.gwt.ui/src/main/resources/com/sap/sailing/gwt/ui/*.gwt.xml; do
+		    mv -v $i.bak $i
+	    done
+    fi
+
 	echo "Build complete. Do not forget to install product..."
 fi
 
 if [[ "$@" == "install" ]] || [[ "$@" == "all" ]]; then
 
-    read -s -n1 -p "Currently branch $active_branch is active and I will deploy to $ACDIR. Do you want to proceed with $@ (y/N): " answer
-    case $answer in
-    "Y" | "y") echo "Continuing";;
-    *) echo "Aborting..."
-       exit;;
-    esac
+	if [ $no_confirmation -eq 0 ]; then
+        read -s -n1 -p "Currently branch $active_branch is active and I will deploy to $ACDIR. Do you want to proceed with $@ (y/N): " answer
+        case $answer in
+        "Y" | "y") echo "Continuing";;
+        *) echo "Aborting..."
+           exit;;
+        esac
+    fi
 
     if [ ! -d $ACDIR ]; then
         echo "Could not find directory $ACDIR - perhaps you are on a wrong branch?"
@@ -344,11 +383,14 @@ if [[ "$@" == "install" ]] || [[ "$@" == "all" ]]; then
     rm -rf $ACDIR/org.eclipse.*
     rm -rf $ACDIR/configuration/org.eclipse.*
 
+    # always overwrite start and stop scripts as they
+    # should never contain any custom logic
+    cp -v $PROJECT_HOME/java/target/start $ACDIR/
+    cp -v $PROJECT_HOME/java/target/stop $ACDIR/
+    cp -v $PROJECT_HOME/java/target/status $ACDIR/
+
     if [ ! -f "$ACDIR/env.sh" ]; then
         cp -v $PROJECT_HOME/java/target/env.sh $ACDIR/
-        cp -v $PROJECT_HOME/java/target/start $ACDIR/
-        cp -v $PROJECT_HOME/java/target/stop $ACDIR/
-        cp -v $PROJECT_HOME/java/target/status $ACDIR/
     fi
 
     if [ ! -f $ACDIR/no-overwrite ]; then
@@ -363,9 +405,6 @@ if [[ "$@" == "install" ]] || [[ "$@" == "all" ]]; then
         cp -v $PROJECT_HOME/configuration/mongodb.cfg $ACDIR/
 
         cp -v $PROJECT_HOME/java/target/env.sh $ACDIR/
-        cp -v $PROJECT_HOME/java/target/start $ACDIR/
-        cp -v $PROJECT_HOME/java/target/stop $ACDIR/
-        cp -v $PROJECT_HOME/java/target/status $ACDIR/
         cp -v $PROJECT_HOME/java/target/udpmirror $ACDIR/
 
         cp -v $PROJECT_HOME/java/target/http2udpmirror $ACDIR
@@ -427,12 +466,14 @@ if [[ "$@" == "remote-deploy" ]]; then
     REMOTE_HOME=`ssh $REMOTE_SERVER_LOGIN 'echo $HOME/servers'`
     REMOTE_SERVER="$REMOTE_HOME/$SERVER"
 
-    read -s -n1 -p "I will deploy the current GIT branch to $REMOTE_SERVER_LOGIN:$REMOTE_SERVER. Is this correct (y/n)? " answer
-    case $answer in
-    "Y" | "y") OK=1;;
-    *) echo "Aborting... nothing has been changed on remote server!"
-    exit;;
-    esac
+	if [ $no_confirmation -eq 0 ]; then
+        read -s -n1 -p "I will deploy the current GIT branch to $REMOTE_SERVER_LOGIN:$REMOTE_SERVER. Is this correct (y/n)? " answer
+        case $answer in
+        "Y" | "y") OK=1;;
+        *) echo "Aborting... nothing has been changed on remote server!"
+        exit;;
+        esac
+    fi
 
     $SSH_CMD "test -d $REMOTE_SERVER/plugins"
     if [[ $? -eq 1 ]]; then
@@ -474,18 +515,20 @@ if [[ "$@" == "remote-deploy" ]]; then
 
     echo "Deployed successfully. I did NOT change any configuration (no env.sh or config.ini or jetty.xml adaption), only code!"
 
-    read -s -n1 -p "Do you want me to restart the remote server (y/n)? " answer
-    case $answer in
-    "Y" | "y") OK=1;;
-    *) echo "Aborting... deployment should be ready by now!"
-    exit;;
-    esac
+	if [ $no_confirmation -eq 0 ]; then
+        read -s -n1 -p "Do you want me to restart the remote server (y/n)? " answer
+        case $answer in
+        "Y" | "y") OK=1;;
+        *) echo "Aborting... deployment should be ready by now!"
+        exit;;
+        esac
 
-    echo ""
-    $SSH_CMD "cd $REMOTE_SERVER && bash -l -c $REMOTE_SERVER/stop"
-    $SSH_CMD "cd $REMOTE_SERVER && bash -l -c $REMOTE_SERVER/start"
+        echo ""
+        $SSH_CMD "cd $REMOTE_SERVER && bash -l -c $REMOTE_SERVER/stop"
+        $SSH_CMD "cd $REMOTE_SERVER && bash -l -c $REMOTE_SERVER/start"
 
-    echo "Restarted remote server. Please check."
+        echo "Restarted remote server. Please check."
+    fi
 fi
 
 if [[ "$@" == "deploy-startpage" ]]; then
